+++ spec/dummy/db/migrate/20150213234410_add_special_price_to_products.rb
@@ -1,4 +1,4 @@
-class AddSpecialPriceToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddSpecialPriceToProducts < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :special_price_cents, :integer
   end