+++ spec/dummy/db/migrate/20150303222230_add_skip_validation_price_cents_to_products.rb
@@ -1,4 +1,4 @@
-class AddSkipValidationPriceCentsToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddSkipValidationPriceCentsToProducts < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :skip_validation_price_cents, :string
   end