+++ lib/money-rails/hooks.rb
@@ -8,30 +8,13 @@ module MoneyRails
         require 'money-rails/active_model/validator'
         require 'money-rails/active_record/monetizable'
         ::ActiveRecord::Base.send :include, MoneyRails::ActiveRecord::Monetizable
-        if defined?(::ActiveRecord) && defined?(::ActiveRecord::VERSION)
-          if ::ActiveRecord::VERSION::MAJOR >= 4
-            rails42               = case
-                                    when ::ActiveRecord::VERSION::MAJOR < 5 && ::ActiveRecord::VERSION::MINOR >= 2
-                                      true
-                                    when ::ActiveRecord::VERSION::MAJOR >= 5
-                                      true
-                                    else
-                                      false
-                                    end
 
-            current_adapter = if ::ActiveRecord::Base.respond_to?(:connection_db_config)
-                                ::ActiveRecord::Base.connection_db_config.configuration_hash[:adapter]
-                              else
-                                ::ActiveRecord::Base.connection_config[:adapter]
-                              end
-
-            postgresql_with_money = rails42 && PG_ADAPTERS.include?(current_adapter)
-          end
-        end
+        current_adapter = ::ActiveRecord::Base.connection_db_config.configuration_hash[:adapter]
+        postgresql_with_money = PG_ADAPTERS.include?(current_adapter)
 
         require "money-rails/active_record/migration_extensions/options_extractor"
         %w{schema_statements table}.each do |file|
-          require "money-rails/active_record/migration_extensions/#{file}_pg_rails4"
+          require "money-rails/active_record/migration_extensions/#{file}_pg"
           if !postgresql_with_money
             require "money-rails/active_record/migration_extensions/#{file}"
           end
@@ -44,11 +27,7 @@ module MoneyRails
       # For Mongoid
       begin; require 'mongoid'; require 'mongoid/version'; rescue LoadError; end
       if defined? ::Mongoid
-        if ::Mongoid::VERSION =~ /^2(.*)/
-          require 'money-rails/mongoid/two' # Loading the file is enough
-        else
-          require 'money-rails/mongoid/money'
-        end
+        require 'money-rails/mongoid/money'
       end
 
       # For ActionView