+++ spec/dummy/db/migrate/20120607210247_add_column_that_allows_nil.rb
@@ -1,4 +1,4 @@
-class AddColumnThatAllowsNil < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddColumnThatAllowsNil < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :optional_price_cents, :integer
   end