+++ spec/active_record/monetizable_spec.rb
@@ -1,5 +1,3 @@
-# encoding: utf-8
-
 require 'spec_helper'
 
 class Sub < Product; end
@@ -20,16 +18,7 @@ if defined? ActiveRecord
         Service.create(charge_cents: 2000, discount_cents: 120)
       end
 
-      def update_product(*attributes)
-        if defined?(::ActiveRecord::VERSION) && ::ActiveRecord::VERSION::MAJOR >= 5
-          product.update(*attributes)
-        else
-          product.update_attributes(*attributes)
-        end
-      end
-
       context ".monetized_attributes" do
-
         it "adds methods to the inheritance chain" do
           class MyProduct < ActiveRecord::Base
             self.table_name = :products
@@ -97,8 +86,8 @@ if defined? ActiveRecord
         expect(product.price_cents).to eq(3210)
       end
 
-      it "correctly updates from a Money object using update_attributes" do
-        expect(update_product(price: Money.new(215, "USD"))).to be_truthy
+      it "correctly updates from a Money object using update" do
+        expect(product.update(price: Money.new(215, "USD"))).to be_truthy
         expect(product.price_cents).to eq(215)
       end
 
@@ -120,6 +109,14 @@ if defined? ActiveRecord
 
         expect(transaction.amount).to eq(Money.new(20000, 'CLP'))
         expect(transaction.amount_cents).to eq(20000)
+      end
+
+      it "update to instance currency field gets applied to converted methods" do
+        transaction = Transaction.create(amount: '200', tax: '10', currency: 'USD')
+        expect(transaction.total).to eq(Money.new(21000, 'USD'))
+
+        transaction.currency = 'CLP'
+        expect(transaction.total).to eq(Money.new(210, 'CLP'))
       end
 
       it "raises an error if trying to create two attributes with the same name" do
@@ -211,9 +208,9 @@ if defined? ActiveRecord
         end
       end
 
-      it "respects numericality validation when using update_attributes" do
-        expect(update_product(price_cents: "some text")).to be_falsey
-        expect(update_product(price_cents: 2000)).to be_truthy
+      it "respects numericality validation when using update" do
+        expect(product.update(price_cents: "some text")).to be_falsey
+        expect(product.update(price_cents: 2000)).to be_truthy
       end
 
       it "uses numericality validation on money attribute" do
@@ -439,9 +436,9 @@ if defined? ActiveRecord
         expect(product.save).to be_truthy
       end
 
-      it "respects numericality validation when using update_attributes on money attribute" do
-        expect(update_product(price: "some text")).to be_falsey
-        expect(update_product(price: Money.new(320, 'USD'))).to be_truthy
+      it "respects numericality validation when using update on money attribute" do
+        expect(product.update(price: "some text")).to be_falsey
+        expect(product.update(price: Money.new(320, 'USD'))).to be_truthy
       end
 
       it "uses i18n currency format when validating" do
@@ -632,7 +629,6 @@ if defined? ActiveRecord
 
       context "when the monetized field is an aliased attribute" do
         it "writes the subunits to the original (unaliased) column" do
-          pending if Rails::VERSION::MAJOR < 4
           product.renamed = "$10.00"
           expect(product.aliased_cents).to eq 10_00
         end
@@ -964,6 +960,25 @@ if defined? ActiveRecord
 
         expect(price).to be_an_instance_of(Money)
         expect(price.amount).not_to eq(value.amount)
+      end
+
+      context 'without a default currency' do
+        let(:product) { OtherProduct.new }
+
+        around do |example|
+          default_currency = Money.default_currency
+          Money.default_currency = nil
+
+          example.run
+
+          Money.default_currency = default_currency
+        end
+
+        it "errors a NoCurrency Error" do
+          expect do
+            product.write_monetized :price, :price_cents, 10.5, false, nil, {}
+          end.to raise_error(Money::Currency::NoCurrency)
+        end
       end
 
       describe "instance_currency_name" do