+++ spec/dummy/db/migrate/20120524052716_create_services.rb
@@ -1,4 +1,4 @@
-class CreateServices < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class CreateServices < ActiveRecord::Migration[7.0]
   def change
     create_table :services do |t|
       t.integer :charge_cents