+++ spec/dummy/db/migrate/20130124023419_add_price_in_a_range_cents_to_products.rb
@@ -1,4 +1,4 @@
-class AddPriceInARangeCentsToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddPriceInARangeCentsToProducts < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :price_in_a_range_cents, :integer
   end