+++ spec/dummy/db/migrate/20140110194016_add_validates_method_amount_cents_to_products.rb
@@ -1,4 +1,4 @@
-class AddValidatesMethodAmountCentsToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddValidatesMethodAmountCentsToProducts < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :validates_method_amount_cents, :integer
   end