+++ spec/dummy/db/migrate/20150107061030_add_delivery_fee_cents_and_restock_fee_cents_to_product.rb
@@ -1,4 +1,4 @@
-class AddDeliveryFeeCentsAndRestockFeeCentsToProduct < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddDeliveryFeeCentsAndRestockFeeCentsToProduct < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :delivery_fee_cents, :integer
     add_column :products, :restock_fee_cents, :integer