+++ spec/dummy/db/migrate/20151026220420_add_optional_amount_to_transactions.rb
@@ -1,4 +1,4 @@
-class AddOptionalAmountToTransactions < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
+class AddOptionalAmountToTransactions < ActiveRecord::Migration[7.0]
   def change
     add_column :transactions, :optional_amount_cents, :integer
   end