+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 - **Breaking change**: Drop support for Rails < 7.0
 
+- Fix `MoneyRails::ActionViewExtension` not available during eager loading (GH-614)
+
 ## 2.0.0
 
 - **Breaking change**: Requires `money` gem version ~> 7.0.